Correciones al backend

+++ src/actions/admin/crudActions.ts
@@ -2,7 +2,10 @@
 import axios from "axios";
 
 export const getDataByid = async (endpoint: string, id: string) => {
-  const urlApi = `${process.env.NESTJS_BACKEND}/${endpoint}/${id}`;
+  const urlBackend = process.env.NETJS_BACKEND;
+
+  if (!urlBackend) throw new Error("Error on construct url backend");
+  const urlApi = `${urlBackend}/${endpoint}/${id}`;
 
   const res = await fetch(urlApi, {
     cache: "no-cache",
@@ -14,7 +17,10 @@ export const getDataByid = async (endpoint: string, id: string) => {
 };
 
 export const createData = async <T>(data: T, endpoint: string, id: string) => {
-  const urlApi = `${process.env.NESTJS_BACKEND}/${endpoint}/${id}`;
+  const urlBackend = process.env.NETJS_BACKEND;
+
+  if (!urlBackend) throw new Error("Error on construct url backend");
+  const urlApi = `${urlBackend}/${endpoint}/${id}`;
 
   try {
     const response = await axios.post<T>(urlApi, data);
@@ -25,7 +31,10 @@ export const createData = async <T>(data: T, endpoint: string, id: string) => {
 };
 
 export const updateData = async <T>(data: T, endpoint: string, id: string) => {
-  const urlApi = `${process.env.NESTJS_BACKEND}/${endpoint}/${id}`;
+  const urlBackend = process.env.NETJS_BACKEND;
+
+  if (!urlBackend) throw new Error("Error on construct url backend");
+  const urlApi = `${urlBackend}/${endpoint}/${id}`;
 
   try {
     const response = await axios.patch<T>(urlApi, data);
@@ -36,7 +45,10 @@ export const updateData = async <T>(data: T, endpoint: string, id: string) => {
 };
 
 export const deleteData = async (endpoint: string, id: string) => {
-  const urlApi = `${process.env.NESTJS_BACKEND}/${endpoint}/${id}`;
+  const urlBackend = process.env.NETJS_BACKEND;
+
+  if (!urlBackend) throw new Error("Error on construct url backend");
+  const urlApi = `${urlBackend}/${endpoint}/${id}`;
 
   try {
     const response = await axios.delete(urlApi);
